# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EMPRENDE",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
...
        "/api/v1/superadmin": {
            "get": {
                "tags": [
                    "Usuario-Superadmin"
                ],
                "summary": "Mostrar información de usuarios superadministradores",
                "operationId": "mostrarSuperadmins",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los emprendimientos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/v1/admin": {
            "get": {
                "tags": [
                    "Usuario-Admin"
                ],
                "summary": "Mostrar información de usuarios administradores",
                "operationId": "mostrarAdmins",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los emprendimientos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/v1/superadmin/create": {
            "post": {
                "tags": [
                    "Usuario-Superadmin"
                ],
                "summary": "Crear información de superadministrador",
                "description": "Crear información de superadministrador",
                "operationId": "crearSuperadmin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "personal_phone": {
                                        "type": "integer"
                                    },
                                    "linkedin": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/admin/create": {
            "post": {
                "tags": [
                    "Usuario-Admin"
                ],
                "summary": "Crear información de administrador",
                "description": "Crear información de administrador",
                "operationId": "crearAdmin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "personal_phone": {
                                        "type": "integer"
                                    },
                                    "linkedin": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/superadmin/{superadmin}": {
            "get": {
                "tags": [
                    "Usuario-Superadmin"
                ],
                "summary": "Mostrar información de un superadministrador en específico",
                "operationId": "mostrarSuperadmin",
                "parameters": [
                    {
                        "name": "superadmin",
                        "in": "path",
                        "description": "id de superadministrador",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/admin/{admin}": {
            "get": {
                "tags": [
                    "Usuario-Admin"
                ],
                "summary": "Mostrar información de un administrador en específico",
                "operationId": "mostrarAdmin",
                "parameters": [
                    {
                        "name": "admin",
                        "in": "path",
                        "description": "id de administrador",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/superadmin/{superadmin}/update": {
            "post": {
                "tags": [
                    "Usuario-Superadmin"
                ],
                "summary": "Actualizar información de superadministrador",
                "description": "Actualizar información de superadministrador",
                "operationId": "actualizarSuperadmin",
                "parameters": [
                    {
                        "name": "superadmin",
                        "in": "path",
                        "description": "id de superadministrador",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "personal_phone": {
                                        "type": "integer"
                                    },
                                    "linkedin": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/v1/superadmin/{superadmin}/destroy": {
            "get": {
                "tags": [
                    "Usuario-Superadmin"
                ],
                "summary": "Desactivar información de superadministrador",
                "description": "Desactivar información de superadministrador",
                "operationId": "desactivarSuperadmin",
                "parameters": [
                    {
                        "name": "superadmin",
                        "in": "path",
                        "description": "id de superadministrador",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "name": "Authorization",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://backend-tesis.test/oauth/authorize",
                        "tokenUrl": "http://backend-tesis.test/oauth/token",
                        "refreshUrl": "http://backend-tesis.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        {
            "oauth2_security_example": [
                "read",
                "write"
            ],
            "passport": []
        }
    ]
}